Modularize analysis
