+++ covid_moonshot/analysis/free_energy.py
@@ -3,7 +3,7 @@ from typing import List
 import numpy as np
 from pymbar import BAR
 from pymbar.mbar import MBAR
-from .core import Binding, PhaseAnalysis, RunAnalysis, Work
+from covid_moonshot.core import PhaseAnalysis, RunAnalysis, Work
 
 
 def mask_outliers(a: np.array, max_value: float, n_devs: float) -> np.array:
@@ -104,28 +104,4 @@ def get_phase_analysis(works: List[Work], min_num_work_values=10) -> PhaseAnalys
         ddelta_f=ddelta_f,
         bar_overlap=bar_overlap,
         num_work_values=len(ws),
-    )
-
-
-def get_run_analysis(
-    complex_works: List[Work], solvent_works: List[Work],
-) -> RunAnalysis:
-
-    try:
-        complex_phase = get_phase_analysis(complex_works)
-    except ValueError as e:
-        raise ValueError(f"Failed to analyze complex: {e}")
-
-    try:
-        solvent_phase = get_phase_analysis(solvent_works)
-    except ValueError as e:
-        raise ValueError(f"Failed to analyze solvent: {e}")
-
-    binding = Binding(
-        delta_f=solvent_phase.delta_f - complex_phase.delta_f,
-        ddelta_f=(complex_phase.ddelta_f ** 2 + solvent_phase.ddelta_f ** 2) ** 0.5,
-    )
-
-    return RunAnalysis(
-        complex_phase=complex_phase, solvent_phase=solvent_phase, binding=binding
     )
